+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Sign releases.
 
+### Changed
+- Release to atlassian-public maven repository.
+
 ## [0.3.0] - 2018-08-27
 [0.3.0]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/release-0.3.0%0Drelease-0.2.0